+++ lib/ast_transform.rb
@@ -4,6 +4,7 @@ require "ast_transform/version"
 require "ast_transform/instruction_sequence"
 require "ast_transform/instruction_sequence/mixin"
 require "ast_transform/instruction_sequence/bootsnap_mixin"
+require "ast_transform/source_parser"
 
 module ASTTransform
   DEFAULT_OUTPUT_PATH = Pathname.new("").join("tmp", "ast_transform").to_s